phase 58.E: fix IF-191 long_source regalloc (highestRegisterNumber uint8 truncation)
Assembler::highestRegisterNumber() truncated the max register index to
uint8_t, silently mapping reg 256 -> 0, 257 -> 1, etc.  The orig keeps
the full int value (the AND/OR sequence at 0x2901b3..0x2901db is a
value-preserving identity).

Once GlobalResources::regNumber crossed 256 within a single
optimization pass (~128 setUserReg calls), removeUnusedRegs +
maxRegister under-reported numRegs to registerAllocation, and high
vregs were never coalesced.  R256+ references survived to the
assembler emit pass, which then rejected them as 'register does not
exist'.

Bisect: 127 setUserReg calls = OK, 128 = fail.  Fix: cast to uint32_t.

Tests:
  - long_source_{hdawg,shfsg}: byte-identical
  - main suite: 1600/1600
  - stress suite: 444/444

+++ reconstructed/src/asm/assembler.cpp
@@ -298,7 +298,15 @@ int64_t Assembler::highestRegisterNumber() const {
 
     if (!found)
         return 0;
-    return (1LL << 32) | static_cast<uint8_t>(maxReg);
+    // IF-191 fix (Phase 58.E): the orig keeps the FULL int value of the
+    // max register (the low-byte/high-bits AND/OR sequence at 0x2901b3..
+    // 0x2901db is a value-preserving identity).  Earlier reconstruction
+    // truncated to uint8_t, which silently mapped reg 256 → 0, reg 257 → 1,
+    // etc.  This caused regalloc / removeUnusedRegs to under-report
+    // numRegs once GlobalResources::regNumber crossed 256 within a single
+    // optimization pass, leaving R256+ references untouched in the output
+    // and triggering "register does not exist" at assemble time.
+    return (1LL << 32) | static_cast<uint32_t>(maxReg);
 }
 
 // 0x28ebd0 — Produce disassembly string.
